refactor: converting Bitmex to typescript
